requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==1.26.15
